+++ public/js/util/dom-lib.js
@@ -1,58 +1,56 @@
-export const getElementsByClassName = (
-  className,
-  base = document.children[0]
-) => {
-  const recursive = (node) => {
+export const getElementsByClassName = (className, base = document.body) => {
+  const searchNode = (node) => {
     const result = [];
-
     if (node.classList.contains(className)) {
       result.push(node);
     }
 
-    for (let i = 0; i < node.children.length; i++) {
-      const childNode = node.children[i];
-      result.push(...recursive(childNode));
+    for (const child of node.children) {
+      result.push(...searchNode(child));
     }
 
     return result;
   };
 
-  return recursive(base);
+  return searchNode(base);
 };
+const as = getElementsByClassName("map-cell");
+console.log(as);
 
-export const getElementByClassName = (
-  className,
-  base = document.children[0]
-) => {
-  const recursive = (node) => {
+export const getElementByClassName = (className, base = document.body) => {
+  let result;
+  const searchNode = (node) => {
     if (node.classList.contains(className)) {
       return node;
-    } else {
-      for (let i = 0; i < node.children.length; i++) {
-        const childNode = node.children[i];
-        const result = recursive(childNode);
+    }
 
-        if (result) return result;
-      }
+    for (let i = 0; i < node.children.length; i++) {
+      const childNode = node.children[i];
+      const result = searchNode(childNode);
+
+      if (result) return result;
     }
   };
 
-  return recursive(base);
+  return searchNode(base);
 };
 
-export const getElementById = (idName, base = document.children[0]) => {
-  const recursive = (node) => {
+const abc = getElementByClassName("map-cell");
+console.log(abc);
+
+export const getElementById = (idName, base = document.body) => {
+  const searchNode = (node) => {
     if (node.id === idName) {
       return node;
     } else {
       for (let i = 0; i < node.children.length; i++) {
         const childNode = node.children[i];
-        const result = recursive(childNode);
+        const result = searchNode(childNode);
 
         if (result) return result;
       }
     }
   };
 
-  return recursive(base);
+  return searchNode(base);
 };
